chore: add CWE ids to findings CSV
Support added in both `findings_csv` and `findings_to_defectdojo`.

This information is useful to obtain in some cases, either to import
into another security intelligence platform or to use as part of the
vulnerability analysis.

For optimization purposes, a short "in-memory" cache is set to avoid
spamming the API looking for vulnerability definition in both commands.

Also rewrote some of the logic to use Python's dictionary `.get()` with
defaults instead of writing if statements, where applicable. Same result,
simpler to read.